requirements.txt
openai
pdfminer.six
python-dotenv 